Move client-side dev JS to dev folder
