+++ packages/next/client/dev/amp-dev.js
@@ -1,7 +1,7 @@
 /* globals __webpack_hash__ */
 import fetch from 'unfetch'
-import EventSourcePolyfill from './event-source-polyfill'
-import { getEventSourceWrapper } from './dev-error-overlay/eventsource'
+import EventSourcePolyfill from '../event-source-polyfill'
+import { getEventSourceWrapper } from './error-overlay/eventsource'
 import { setupPing } from './on-demand-entries-utils'
 
 if (!window.EventSource) {
