+++ packages/next/client/dev/error-overlay/eventsource.js
