+++ packages/next/client/dev/error-overlay/format-webpack-messages.d.ts
