+++ packages/next/client/next-dev.js
@@ -1,8 +1,8 @@
 import initNext, * as next from './'
 import EventSourcePolyfill from './event-source-polyfill'
-import initOnDemandEntries from './on-demand-entries-client'
-import initWebpackHMR from './webpack-hot-middleware-client'
-import initializeBuildWatcher from './dev-build-watcher'
+import initOnDemandEntries from './dev/on-demand-entries-client'
+import initWebpackHMR from './dev/webpack-hot-middleware-client'
+import initializeBuildWatcher from './dev/dev-build-watcher'
 
 // Temporary workaround for the issue described here:
 // https://github.com/zeit/next.js/issues/3775#issuecomment-407438123
